REF: move pytmc.epics->pytmc.linter
