+++ pytmc/__init__.py
@@ -4,7 +4,7 @@ from ._version import get_versions
 from .record import RecordPackage, EPICSRecord
 from . import pragmas
 from . import parser
-from . import epics
+from . import linter
 
 logger = logging.getLogger(__name__)
 __version__ = get_versions()['version']
@@ -14,7 +14,7 @@ del get_versions
 __all__ = [
     'EPICSRecord',
     'RecordPackage',
-    'epics',
+    'linter',
     'logger',
     'parser',
     'pragmas',
